+++ src/main/java/frc/robot/commands/autonomous/TrajectoryFollower.java
@@ -2,8 +2,8 @@ package frc.robot.commands.autonomous;
 
 import frc.robot.Constants.TrajectoryConstants;
 import frc.robot.Constants.DriveConstants;
-import frc.robot.subsystems.Drivebase;
-import frc.robot.subsystems.togglesystem.ToggleGeneralMotors;
+import frc.robot.subsystems.dreadsubsystem.Drivebase;
+import frc.robot.subsystems.dreadsubsystem.Drivebase;
 import edu.wpi.first.math.controller.PIDController;
 import edu.wpi.first.math.controller.RamseteController;
 import edu.wpi.first.math.controller.SimpleMotorFeedforward;
