+++ src/main/java/frc/robot/commands/mastertoggle/Shoot.java
@@ -1,15 +1,16 @@
-package frc.robot.commands.subtoggle;
+package frc.robot.commands.mastertoggle;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.togglesystem.ToggleGeneralMotors;
-import frc.robot.subsystems.togglesystem.Turret;
+import frc.robot.subsystems.dreadsubsystem.Turret;
+
+import frc.robot.subsystems.dreadsubsystem.MasterSubsystem;
 
 public class Shoot extends CommandBase {
     private final Turret m_shooter;
-    private final ToggleGeneralMotors m_index;
+    private final MasterSubsystem m_index;
     private boolean m_isDone;
 
-    public Shoot(Turret shooter, ToggleGeneralMotors index) {
+    public Shoot(Turret shooter, MasterSubsystem index) {
         m_shooter = shooter;
         m_index = index;
         m_isDone = false;
