+++ src/main/java/frc/robot/commands/mastertoggle/dstoggle/ToggleGear.java
@@ -1,13 +1,13 @@
-package frc.robot.commands.dstoggle;
+package frc.robot.commands.mastertoggle.dstoggle;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.pneumatics.SolenoidsSystem;
+import frc.robot.subsystems.dreadsubsystem.MasterSubsystem;
 
 public class ToggleGear extends CommandBase {
-  private final SolenoidsSystem m_solenoids;
+  private final MasterSubsystem m_solenoids;
   private boolean m_isDone;
 
-  public ToggleGear(SolenoidsSystem subsystem) {
+  public ToggleGear(MasterSubsystem subsystem) {
     m_solenoids = subsystem;
     m_isDone = false;
     addRequirements(subsystem);
